chore(lint): sweep ruff debt + version-agnostic packaging tests
22 pre-existing ruff errors surfaced during the v0.1.4 release workflow's
fast-suite gate. Most were stylistic (en-dash unicode in docstrings,
unnecessary int() casts, regex metacharacters in pytest.raises patterns,
unused unpacked variable). One was a real bug: tests/test_parity.py used
`mostlyright.research()` but never imported the module. The test is marked
@pytest.mark.live so it only runs locally; the missing import was hidden
because the live tests never ran in CI.

Changes:
- tests/test_parity.py: add `import mostlyright`
- pyproject.toml: per-file F821 ignores for the 2 fixture scripts that
  lazy-import mostlyright inside main() (capture_drift.py + measure_ulp_drift.py)
- packages/markets/src/mostlyright/markets/kalshi_trades.py: en-dash → hyphen
- packages/markets/tests/test_kalshi_trades.py: en-dash in comment → hyphen
- packages/weather/src/mostlyright/weather/_fetchers/_iem_mos.py: drop
  unnecessary `int(round(...))` → `round(...)` (Python 3 round() returns int)
- packages/core/src/mostlyright/_compose.py: `short_kalshi` → `_short_kalshi`
- packages/markets/src/mostlyright/markets/_kalshi_client.py: SIM201 fix
  `not (s == s)` → `s != s` (canonical NaN check)
- 3 pytest.raises patterns made raw strings + escaped (test_compose.py,
  test_iem.py, test_iem_catalog_fetch_forecasts.py)

Plus tests/test_packaging.py + tests/test_wheel_layout.py: convert the
version-pinned tests from "==0.1.0rc1" assertions to version-agnostic
lockstep + 0.1.x-range assertions. Originally written at Phase 4 PKG-01
for the rc1 era; never updated as the project versioned forward through
0.1.0..0.1.4. Now they survive any 0.1.x bump without modification (and
will need updating at 0.2.0 / 1.0.0).

uv run ruff check .: All checks passed!
uv run pytest -m "not live" -q: all green.

+++ packages/weather/src/mostlyright/weather/_fetchers/_iem_mos.py
@@ -144,7 +144,7 @@ def _parse_mos_row(
     valid_at = _parse_dt(row.get("ftime") or row.get("valid_time"))
     if issued_at is None or valid_at is None:
         return None
-    forecast_hour = int(round((valid_at - issued_at).total_seconds() / 3600))
+    forecast_hour = round((valid_at - issued_at).total_seconds() / 3600)
     return {
         "station": station,
         "issued_at": issued_at,
@@ -154,7 +154,7 @@ def _parse_mos_row(
         "temp_c": _f_to_c(_maybe_float(row, "tmp")),
         "dew_point_c": _f_to_c(_maybe_float(row, "dpt")),
         "wind_speed_ms": _kt_to_ms(_maybe_float(row, "wsp")),
-        "wind_dir_deg": (int(round(d)) if (d := _maybe_float(row, "wdr")) is not None else None),
+        "wind_dir_deg": (round(d) if (d := _maybe_float(row, "wdr")) is not None else None),
         "precip_probability": _pct_to_unit(_maybe_float(row, "pop12")),
         "sky_cover_pct": None,  # IEM MOS doesn't expose a single sky-cover %
         "source": "iem.archive",
